Empty files to hold the MultiSampler class
the multisampler class willi allow the use of a single PRNG to sample an
integer (e.g. to randomly obtain indices) sample a uniform distribution
from 0-1 to check acceptance of points with lower likelihoods, and to sample
from one or more specialized probability distributions (allowing construction
of new proposal points and the like).

Signed-off-by: James Till Matta <jamesmatta@gmail.com>